# Fix performance regression in dplyr 0.8.2