@romainfrancois 2019-07-03
The bug was introduced here: https://github.com/tidyverse/dplyr/commit/036de90fbf9e3eef72c015982a5d1294d2157a2c#r34165423

where for some reason, the materialized binding was no longer defined in the resolved environment, which is the entire point of how the dplyr data mask layout works internally 😬